moved callback interface
added comments. app description.
pick a fragment. get to work.
