+++ app/src/main/java/acme/android/finalapp/MainActivity.kt
@@ -1,8 +1,10 @@
 package acme.android.finalapp
 
+import acme.android.finalapp.helper.FragmentListener
+import acme.android.finalapp.ui.BlankFragment
+import acme.android.finalapp.ui.HomeFragment
 import android.os.Bundle
 import android.util.Log
-import com.google.android.material.snackbar.Snackbar
 import androidx.appcompat.app.AppCompatActivity
 import android.view.Menu
 import android.view.MenuItem
@@ -10,42 +12,34 @@ import androidx.fragment.app.Fragment
 
 import kotlinx.android.synthetic.main.activity_main.*
 
-class MainActivity : BlankFragment.OnFragmentInteractionListener, AppCompatActivity() {
+class MainActivity : FragmentListener, AppCompatActivity() {
+    private var showfmenu: Boolean = false
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        setSupportActionBar(toolbar)
 
-        fab.setOnClickListener { view ->
-           if(!showfmenu) showFM()
-             else hideFM()
-        }
-
-        showFrag(0) //show the "home fragment"
-        iniFM() //setup floating button menu
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu): Boolean {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        menuInflater.inflate(R.menu.menu_main, menu)
-        return true
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        return when (item.itemId) {
-            R.id.action_settings -> true
-            else -> super.onOptionsItemSelected(item)
-        }
+    private fun log(msg:String){
+        Log.d("Movie APP: ", msg)
     }
 
 
+
+    /*someone please come up with another. more the better.
+
+    pick a fragment. add your name. commit. so we know everyone is online.!
+
+    Fragments:
+    1 Home/Search fragment - Saul
+    2 Favorites Fragment -
+        -save imdb id (shared preference or external file) + download poster
+    3 Browse (Abc/genre/year...?) -
+            scraping imdb?
+
+    4 watch trailers? new/popular
+    5   another thing. idk...
+     */
 
     private fun showFrag(op: Int){
 
+        //add new fragment calls here.
         var f : Fragment = HomeFragment.newInstance()
         when(op){
             0,2 -> f = BlankFragment.newInstance("", "") //replace with real fragments
@@ -57,13 +51,25 @@ class MainActivity : BlankFragment.OnFragmentInteractionListener, AppCompatActiv
             R.anim.slideout_down).replace(R.id.fragmentFrame,f).commit()
 
     }
-    private var showfmenu: Boolean = false
+    
 
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        setSupportActionBar(toolbar)
 
-    private fun log(msg:String){
-        Log.d("Movie APP: ", msg)
+        fab.setOnClickListener { view ->
+            if(!showfmenu) showFM()
+            else hideFM()
+        }
+
+        showFrag(0) //show the "home fragment"
+        iniFM() //setup floating button menu
     }
+    /* ---------- Floating button Menu ------------- */
 
+
+    //main menu listener
     private fun iniFM(){
         fab1.setOnClickListener { showFrag(0) }
         fab1.setOnLongClickListener{
@@ -98,6 +104,15 @@ class MainActivity : BlankFragment.OnFragmentInteractionListener, AppCompatActiv
         fab3.animate().translationY(0f)
 
     }
+
     /* ---------- Floating button Menu ------------- */
 
+
+
+    /* ---------- Fragment callbacks ------------- */
+
+
+    override fun returnSearchedInfo(id: String) {
+        TODO("example call back.") //T
+    }
 }
