+++ app/src/main/java/acme/android/finalapp/ui/BlankFragment.kt
@@ -1,5 +1,7 @@
-package acme.android.finalapp
+package acme.android.finalapp.ui
 
+import acme.android.finalapp.R
+import acme.android.finalapp.helper.FragmentListener
 import android.content.Context
 import android.net.Uri
 import android.os.Bundle
@@ -16,7 +18,7 @@ private const val ARG_PARAM2 = "param2"
 /**
  * A simple [Fragment] subclass.
  * Activities that contain this fragment must implement the
- * [BlankFragment.OnFragmentInteractionListener] interface
+ * [FragmentListener] interface
  * to handle interaction events.
  * Use the [BlankFragment.newInstance] factory method to
  * create an instance of this fragment.
@@ -25,7 +27,7 @@ class BlankFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
-    private var listener: OnFragmentInteractionListener? = null
+    private var listener: FragmentListener? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -43,14 +45,11 @@ class BlankFragment : Fragment() {
         return inflater.inflate(R.layout.fragment_blank, container, false)
     }
 
-    // TODO: Rename method, update argument and hook method into UI event
-    fun onButtonPressed(uri: Uri) {
-       // listener?.onFragmentInteraction(uri)
-    }
+
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
-        if (context is OnFragmentInteractionListener) {
+        if (context is FragmentListener) {
             listener = context
         } else {
             throw RuntimeException(context.toString() + " must implement OnFragmentInteractionListener")
@@ -62,20 +61,7 @@ class BlankFragment : Fragment() {
         listener = null
     }
 
-    /**
-     * This interface must be implemented by activities that contain this
-     * fragment to allow an interaction in this fragment to be communicated
-     * to the activity and potentially other fragments contained in that
-     * activity.
-     *
-     *
-     * See the Android Training lesson [Communicating with Other Fragments]
-     * (http://developer.android.com/training/basics/fragments/communicating.html)
-     * for more information.
-     */
-    interface OnFragmentInteractionListener {
 
-    }
 
     companion object {
         /**
