+++ app/src/main/java/acme/android/finalapp/ui/HomeFragment.kt
@@ -1,25 +1,23 @@
-package acme.android.finalapp
+package acme.android.finalapp.ui
 
+import acme.android.finalapp.R
+import acme.android.finalapp.helper.FragmentListener
 import android.content.Context
-import android.net.Uri
 import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
-import android.os.Message
 import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import kotlinx.android.synthetic.main.fragment_home.*
-import java.io.BufferedInputStream
-import java.net.HttpURLConnection
 import java.net.URL
 import java.util.concurrent.Executors
 
 class HomeFragment : Fragment() {
-    private var listener: BlankFragment.OnFragmentInteractionListener? = null
+    private var listener: FragmentListener? = null
 
+
+    //home fragment show last user search, or a random title.
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -71,10 +69,10 @@ class HomeFragment : Fragment() {
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
-        if (context is BlankFragment.OnFragmentInteractionListener) {
+        if (context is FragmentListener) {
             listener = context
         } else {
-            throw RuntimeException(context.toString() + " must implement OnFragmentInteractionListener")
+            throw RuntimeException(context.toString() + " must implement FragmentListener")
         }
     }
